+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Features
 * Add a new Endpoint type to support unix sockets (PYTHON-1098)
 * DataStax Cloud enablement (PYTHON-1075)
 * Add creds.zip support (PYTHON-1097)
+* Detect CaaS and change consistency default (PYTHON-1145)
 
 3.19.0
 ======